+++ R/ee_download.R
@@ -820,7 +820,7 @@ ee_drive_to_local <- function(task,
     ee_save_credential(pdrive = drive_credential)
     message(
       "Google Drive credentials were not loaded.",
-      " Running ee_Initialize(email = '",ee_user[["email"]],"', drive = TRUE)",
+      " Running ee_Initialize(user = '", ee_user[["email"]], "', drive = TRUE)",
       " to fix."
     )
   }
@@ -846,7 +846,11 @@ ee_drive_to_local <- function(task,
   }
 
   if (public) {
-    files_gd <- googledrive::drive_share_anyone(files_gd, verbose = FALSE)
+    googledrive::with_drive_quiet({
+      files_gd <- googledrive::drive_share_anyone(
+        file = files_gd
+      )
+    })
   }
 
   # (Problem) Google Drive support files with the same name
@@ -924,12 +928,21 @@ ee_drive_to_local <- function(task,
   #   )
   # }
   for (index in seq_len(nrow(to_download))) {
-    googledrive::drive_download(
-      file = to_download[index, ],
-      path = filenames_local[index],
-      overwrite = overwrite,
-      verbose = !quiet
-    )
+    if (!quiet) {
+      googledrive::with_drive_quiet({
+        googledrive::drive_download(
+          file = to_download[index, ],
+          path = filenames_local[index],
+          overwrite = overwrite
+        )
+      })
+    } else {
+      googledrive::drive_download(
+        file = to_download[index, ],
+        path = filenames_local[index],
+        overwrite = overwrite
+      )
+    }
   }
 
   if (metadata) {
@@ -1066,7 +1079,7 @@ ee_gcs_to_local <- function(task,
     ee_save_credential(pgcs = gcs_credential[["path"]])
     message(
       "Google Cloud Storage credentials were not loaded.",
-      " Running ee_Initialize(email = '",ee_user[["email"]],"', gcs = TRUE)",
+      " Running ee_Initialize(user = '", ee_user[["email"]], "', gcs = TRUE)",
       " to fix."
     )
   }