+++ tests/testthat/test-Initialize.R
@@ -45,3 +45,12 @@ test_that("ee_users",{
   counts <- ee_user_info()
   expect_type(counts, "list")
 })
+
+test_that("ee_connect_to_py",{
+  expect_error(rgee:::ee_connect_to_py("demo.py"))
+})
+
+test_that("ee_install_set_init_message", {
+  response <- rgee:::ee_init_message(debug = TRUE)
+  testthat::expect_true(response)
+})