+++ tests/testthat/test-check.R
@@ -12,3 +12,23 @@ test_that("ee_check ",{
   expect_true(ee_check())
 })
 
+
+test_that("ee_wrong_message",{
+  result <- rgee:::ee_wrong_message("sf")
+  expect_null(result)
+})
+
+
+# test_that("ee_clean_credentials",{
+#   result <- rgee:::ee_clean_credentials("user_demo")
+#   expect_true(result)
+# })
+
+
+test_that("ee_install check", {
+  message = ee_install(confirm = FALSE)
+  ee_clean_pyenv()
+  expect_true(message)
+})
+
+